requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2024.1
icalendar>=5.0.12
